example/requirements.txt
Django==2.0
diff-match-patch==20110725.1
django-germanium==2.1.0
six==1.10.0
coverage==4.0.2
pyprind==2.9.9
coveralls==1.1
filemagic==1.6
unidecode==0.4.20
pillow==4.1.1
-e ../
